sanitizer-api/sanitizer-config.tentative.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>

<body>
<script>
    let default_option = {};
    test(t => {
      let s = new Sanitizer();
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, default_option);
    }, "SanitizerAPI creator without config.");

    test(t => {
      let s = new Sanitizer({});
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, default_option);
    }, "SanitizerAPI creator with empty config.");

    test(t => {
      let s = new Sanitizer(null);
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, default_option);
    }, "SanitizerAPI creator with null as config.");

    test(t => {
      let s = new Sanitizer(undefined);
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, default_option);
    }, "SanitizerAPI creator with undefined as config.");

    test(t => {
      let s = new Sanitizer({testConfig: [1,2,3], attr: ["test", "i", "am"]});
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, default_option);
    }, "SanitizerAPI creator with config ignore unknown values.");

    const config_names = ["dropElements", "blockElements", "allowElements", "dropAttributes", "blockAttributes", "allowAttributes"];
    config_names.forEach(cname => {
      test(t => {
        let options = {};
        options[cname] = ["div"];
        let result = {};
        result[cname] = ["DIV"];

        let s = new Sanitizer(options);
        assert_true(s instanceof Sanitizer);
        assert_object_equals(s.creationOptions, result);

        options[cname].push("test");
        assert_object_equals(s.creationOptions, result);

        let new_options = {};
        new_options[cname] = ["test", "t"];
        s.creationOptions = new_options;
        assert_object_equals(s.creationOptions, result);

        s.creationOptions[cname] = [1,2,3];
        assert_object_equals(s.creationOptions, result);
      }, "SanitizerAPI config " + cname + " is not editable.");

    let testcases = [[], null, undefined];
    testcases.forEach(c => {
      let options = {};
      options[cname] = c;
      test(t => {
        let s = new Sanitizer(options);
        assert_true(s instanceof Sanitizer)
        assert_object_equals(s.creationOptions, default_option);
        assert_equals(s.sanitizeToString("<div>balabala<i>test</i></div>"), "<div>balabala<i>test</i></div>");
      }, "SanitizerAPI creator with config " + JSON.stringify(options, function(k,v){return v===undefined?"::undefined::":v}).replace(new RegExp("\"::undefined::\"", 'g'), "undefined") + ".");
    });

    testcases = [123, "div"];
    testcases.forEach(c => {
      options = {};
      options[cname] = c;
      test(t => {
        assert_throws_js(TypeError, _ => {let s = new Sanitizer(options)});
      }, "SanitizerAPI creator with config " + JSON.stringify(options) + ".");
    });
  });

</script>
</body>
</html>
